feat(sync-actions): add assets actions to variants and update tests
