+++ packages/sync-actions/src/products.js
@@ -8,7 +8,6 @@ import type {
 } from 'types/sdk'
 import createBuildActions from './utils/create-build-actions'
 import createMapActionGroup from './utils/create-map-action-group'
-import actionsMapAssets from './assets-actions'
 import * as productActions from './product-actions'
 import * as diffpatcher from './utils/diffpatcher'
 import findMatchingPairs from './utils/find-matching-pairs'
@@ -113,7 +112,7 @@ function createProductMapActions(
 
     allActions.push(
       mapActionGroup('assets', (): Array<UpdateAction> =>
-        actionsMapAssets(diff, oldObj, newObj)
+        productActions.actionsMapAssets(diff, oldObj, newObj)
       )
     )
